Landing Scene, WebGL Convert, Camera Switch
Built a landing page for the game which has the option to implement the experimental generation. Converted the project to a WebGL deployment. Implemented an overview ortographic camera for an a better view of generations. Camera swap button is Space. Escape now leads to the landing scene. Return(Enter) and PadEnter now generate a new level without having to go back to the landing page.